requirements.txt
Flask
faker
kafka-python
flask_apscheduler